feat: agregar nuevas secciones de galería y video, eliminar secciones obsoletas
