--- src/app/[locale]/viajeros/components/GallerySection5.tsx
@@ -1,20 +1,0 @@
-/* eslint-disable @next/next/no-img-element */
-
-import { SectionWrapper } from '@/components/layout/section/base-section';
-
-export const GallerySection5 = () => {
-    return (
-        <SectionWrapper>
-            <div className="container mx-auto">
-                {/* Imagen panorámica grande */}
-                <div className="relative w-full aspect-[16/9] overflow-hidden shadow-md ">
-                    <img
-                        src="/gallery/seccion5/ImgLiving.webp"
-                        alt="interior living"
-                        className="object-cover w-full h-full"
-                    />
-                </div>
-            </div>
-        </SectionWrapper>
-    );
-};
